Check point 3

+++ Backend/user-ms/src/main/java/com/user_ms/controller/UserController.java
@@ -1,8 +1,7 @@
 package com.user_ms.controller;
 
-import com.user_ms.dto.GetUserDto;
-import com.user_ms.dto.PostUserDto;
-import com.user_ms.dto.PutUserDto;
+import com.user_ms.controller.client.TokenValidator;
+import com.user_ms.dto.*;
 import com.user_ms.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.transaction.Transactional;
@@ -14,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
+import java.util.List;
 
 @RestController
 @RequestMapping("/user")
@@ -21,35 +21,53 @@ public class UserController {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private TokenValidator tokenValidator;
+
     @GetMapping("/id")
     @Operation(summary="Retorna usuário com base no id inserido, se existir", description="Retorna usuário com base no id inserido, se existir.")
-    public ResponseEntity<GetUserDto> getUserById(@RequestParam long id){
-        GetUserDto getUserDto = userService.getUserById(id);
+    public ResponseEntity<GetUserDto> getUserById(@RequestHeader("Authorization") String requestHeader,
+                                                  @RequestParam long id){
+        UserHasRoleDto data = new UserHasRoleDto(requestHeader, List.of(Role.ROLE_ADMIN.name(), Role.ROLE_USER.name()));
 
-        if( getUserDto == null )
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if(this.tokenValidator.userHasRole(data).getBody().booleanValue()) {
+            GetUserDto getUserDto = userService.getUserById(id);
 
-        return ResponseEntity.ok(getUserDto);
+            if (getUserDto == null)
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return ResponseEntity.ok(getUserDto);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @GetMapping("/email")
     @Operation(summary="Retorna usuário com base no email inserido, se existir", description="Retorna usuário com base no email inserido, se existir.")
-    public ResponseEntity<GetUserDto> getUserByEmail(@RequestParam String email){
-        GetUserDto getUserDto = userService.getUserByEmail(email);
+    public ResponseEntity<GetUserDto> getUserByEmail(@RequestHeader("Authorization") String requestHeader,
+                                                     @RequestParam String email){
+        UserHasRoleDto data = new UserHasRoleDto(requestHeader, List.of(Role.ROLE_ADMIN.name(), Role.ROLE_USER.name()));
 
-        if( getUserDto == null )
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if(this.tokenValidator.userHasRole(data).getBody().booleanValue()) {
+            GetUserDto getUserDto = userService.getUserByEmail(email);
 
-        return ResponseEntity.ok(getUserDto);
+            if( getUserDto == null )
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return ResponseEntity.ok(getUserDto);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @GetMapping()
     @Operation(summary="Retorna página de usuários", description="Retorna página de usuários.")
-    public ResponseEntity<Page<GetUserDto>> getUsersPaginated(@RequestParam(defaultValue = "0") int page,
+    public ResponseEntity<Page<GetUserDto>> getUsersPaginated(@RequestHeader("Authorization") String requestHeader,
+                                                              @RequestParam(defaultValue = "0") int page,
                                                               @RequestParam(defaultValue = "100") int size){
-        Page<GetUserDto> getUserDtoList = userService.getUsersPaginated(page, size);
+        UserHasRoleDto data = new UserHasRoleDto(requestHeader, List.of(Role.ROLE_ADMIN.name(), Role.ROLE_USER.name()));
 
-        return ResponseEntity.ok(getUserDtoList);
+        if(this.tokenValidator.userHasRole(data).getBody().booleanValue()) {
+            Page<GetUserDto> getUserDtoList = userService.getUsersPaginated(page, size);
+            return ResponseEntity.ok(getUserDtoList);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @PostMapping()
@@ -57,31 +75,41 @@ public class UserController {
     @Operation(summary="Cria usuário", description="Cria usuário.")
     public ResponseEntity<GetUserDto> postUser(@RequestBody PostUserDto postUserDto,
                                                UriComponentsBuilder uriBuilder){
-        GetUserDto getUserDto = this.userService.postUser(postUserDto);
 
+        GetUserDto getUserDto = this.userService.postUser(postUserDto);
         URI uri = uriBuilder.path("/user/id")
                             .buildAndExpand(getUserDto.id())
                             .toUri();
-
         return ResponseEntity.created(uri).body(getUserDto);
+
     }
 
     @PutMapping("/{id}")
     @Transactional
     @Operation(summary="Atualiza usuário", description="Atualiza usuário.")
-    public ResponseEntity<GetUserDto> putUser(@PathVariable long id,
+    public ResponseEntity<GetUserDto> putUser(@RequestHeader("Authorization") String requestHeader,
+                                              @PathVariable long id,
                                               @RequestBody PutUserDto putUserDto){
-        GetUserDto getUserDto = this.userService.putUser(id, putUserDto);
+        UserHasRoleDto data = new UserHasRoleDto(requestHeader, List.of(Role.ROLE_ADMIN.name(), Role.ROLE_USER.name()));
 
-        return new ResponseEntity<>(getUserDto, HttpStatus.OK);
+        if(this.tokenValidator.userHasRole(data).getBody().booleanValue()) {
+            GetUserDto getUserDto = this.userService.putUser(id, putUserDto);
+            return new ResponseEntity<>(getUserDto, HttpStatus.OK);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     @Operation(summary="Deleta usuário", description="Deleta usuário.")
-    public ResponseEntity<GetUserDto> deleteUser(@PathVariable long id){
-        GetUserDto getUserDto = this.userService.deleteUser(id);
+    public ResponseEntity<GetUserDto> deleteUser(@RequestHeader("Authorization") String requestHeader,
+                                                 @PathVariable long id){
+        UserHasRoleDto data = new UserHasRoleDto(requestHeader, List.of(Role.ROLE_ADMIN.name(), Role.ROLE_USER.name()));
 
-        return new ResponseEntity<>(getUserDto, HttpStatus.OK);
+        if(this.tokenValidator.userHasRole(data).getBody().booleanValue()) {
+            GetUserDto getUserDto = this.userService.deleteUser(id);
+            return new ResponseEntity<>(getUserDto, HttpStatus.OK);
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
     }
 }
